swift_project/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
